Migrate to qt6 (#4)
* Migrate accounts and bluez

* Migrated mpris

* Migrated networkmanagement

* Create devcontainer.json

* Delete .devcontainer/devcontainer.json

* Fix an error in appletproxymodel

* Fix screen

* Added Debian experimental to CI

* Fix system

* Fix audio

* Fix error and deprecated methods

* Bump version to 2.0.0-1~dev1

* Change QT_SELECT to 6

* Update debhlper compability to 11

+++ bluez/declarativeadapter.cpp
@@ -9,7 +9,7 @@
 #include "declarativeadapter.h"
 #include "declarativedevice.h"
 
-static int devicesCountDeclarativeAdapterFunction(QQmlListProperty<DeclarativeDevice> *property)
+static long long devicesCountDeclarativeAdapterFunction(QQmlListProperty<DeclarativeDevice> *property)
 {
     Q_ASSERT(qobject_cast<DeclarativeAdapter *>(property->object));
     DeclarativeAdapter *adapter = static_cast<DeclarativeAdapter *>(property->object);
@@ -17,7 +17,7 @@ static int devicesCountDeclarativeAdapterFunction(QQmlListProperty<DeclarativeDe
     return adapter->m_devices.count();
 }
 
-static DeclarativeDevice *devicesAtDeclarativeAdapterFunction(QQmlListProperty<DeclarativeDevice> *property, int index)
+static DeclarativeDevice *devicesAtDeclarativeAdapterFunction(QQmlListProperty<DeclarativeDevice> *property, long long index)
 {
     Q_ASSERT(qobject_cast<DeclarativeAdapter *>(property->object));
     DeclarativeAdapter *adapter = static_cast<DeclarativeAdapter *>(property->object);
